requirements.txt
django
djangorestframework
drf-spectacular
gunicorn
